oslayer/keyboardcontrol: rework classes declaration
Sub-classing the platform KeyboardCapture/KeyboardEmulation class
as keyboardcontrol.KeyboardCapture can lead to non-obvious issues,
e.g. because KeyboardCapture in the context of machine.keyboard
is not the same class as KeyboardCapture in the context of
oslayer.xkeyboardcontrol, making using class attributes more
difficult.

Also move the keys layout to the keyboard machine where it belongs.